requirements.txt
# this file is deprecated and will be removed

ConcurrentLogHandler==0.9.1
Flask==1.0.2  # Updated from 0.12
Jinja2==2.10  # Updated from 2.9.5
MarkupSafe==1.0
PyDispatcher==2.0.5
PyYAML==3.12
Scrapy==1.5.0  # Updated from 1.4.0
Twisted==18.4.0  # Updated from 17.5.0
Werkzeug==0.14.1  # Updated from 0.12.1
attrs==18.1.0  # Updated from 17.2.0
cffi==1.11.5  # Updated from 1.10.0
characteristic==14.3.0
click==6.7
coverage==4.5.1  # Updated from 4.3.4
cryptography==2.3
cssselect==1.0.3  # Updated from 1.0.1
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
idna==2.6
ipaddress==1.0.22  # Updated from 1.0.18
itsdangerous==0.24
jsonschema==2.6.0
kafka-python==1.4.3  # Updated from 1.3.4
kazoo==2.4.0
lxml==4.2.1  # Updated from 3.8.0
mock==2.0.0
nose==1.3.7
parsel==1.4.0  # Updated from 1.2.0
pbr==4.0.3  # Updated from 3.1.1
pyOpenSSL==18.0.0  # Updated from 17.2.0
pyasn1==0.4.3  # Updated from 0.3.2
pyasn1-modules==0.2.1  # Updated from 0.0.11
pycparser==2.18
python-json-logger==0.1.8
python-redis-lock==3.2.0
queuelib==1.5.0  # Updated from 1.4.2
redis>=3.0  # Updated from 2.10.5
requests==2.18.4  # Updated from 2.18.3
requests-file==1.4.3  # Updated from 1.4.2
retrying==1.3.3
./utils # scutils==1.3.0dev0
service-identity==17.0.0
six==1.11.0  # Updated from 1.10.0
testfixtures==6.0.2  # Updated from 5.1.1
tldextract==2.2.0  # Updated from 2.1.0
ujson==1.35
w3lib==1.19.0  # Updated from 1.18.0
zope.interface==4.5.0  # Updated from 4.4.2
pymongo==4.1.1
python-dotenv==0.20.0
pydantic==1.9.1
phantomjs==1.4.1
scrapy-splash==0.8.0

# Generated with piprot 0.9.10
